Implementado registo e login

+++ scriptIndex.js
@@ -1,18 +1,24 @@
 w3.includeHTML(() =>  {
-  // Adiciona o evento de clique ao ícone de hambúrguer para alternar o aside
-  newAccountBtn.addEventListener('click', openRegistry)
-  logoutButton.addEventListener('click', logout);
-  registerBtn.addEventListener('click', register);
-  var form = document.getElementById("edit-id-form");
-  const modalRegister = document.getElementById("modal-register");
 
-  var closeRegister = document.getElementsByClassName("close-id")[0];
-  var logged = sessionStorage.getItem('logged');
+  //Verifica se existe user no localstorage, caso contrário cria um array vazio
+  if(!localStorage.getItem("users")){
+    localStorage.setItem('users', JSON.stringify([]));
+  }
 
+  //Verifica se existe utilizador l
   if(logged === null){
     sessionStorage.setItem('logged', 'false');
   }
   
+  const loginForm = document.getElementById('login-form');
+  const welcomeMessage = document.getElementById('mensagem_boasVindas');
+
+  // Verificar se há um nome de utilizador armazenado no Session Storage
+  const storedUsername = sessionStorage.getItem('username');
+  if (storedUsername) {
+    welcomeMessage.textContent = `Olá, ${storedUsername}!`;
+  } 
+
   //caso não exista um utilizador logged, apresenta o botão de login
   if((sessionStorage.getItem('logged') === 'false')){
     console.log('not a user logged');
@@ -21,16 +27,25 @@ w3.includeHTML(() =>  {
     const loginMessage = document.getElementById("loginMessage");
     loginMessage.style.visibility='hidden';
   }
+
+
+
+
+  // Adiciona o evento de clique ao ícone de hambúrguer para alternar o aside
+  newAccountBtn.addEventListener('click', openRegistry)
+  logoutButton.addEventListener('click', logout);
+  registerBtn.addEventListener('click', register);
+  var form = document.getElementById("edit-id-form");
+  const modalRegister = document.getElementById("modal-register");
+  var closeRegister = document.getElementsByClassName("close-id")[0];
+  var logged = sessionStorage.getItem('logged');
+
   
-    const loginForm = document.getElementById('login-form');
-    const welcomeMessage = document.getElementById('mensagem_boasVindas');
-
-    // Verificar se há um nome de utilizador armazenado no Session Storage
-    const storedUsername = sessionStorage.getItem('username');
-    if (storedUsername) {
-      welcomeMessage.textContent = `Olá, ${storedUsername}!`;
-    } 
-
+  
+  
+  
+    
+    
     const hamburger = document.getElementById('hamburger');
 
     hamburger.addEventListener('click', toggleAside);
@@ -47,15 +62,19 @@ w3.includeHTML(() =>  {
     });
 
     function login(username){
-      console.log('correu o login');
-      sessionStorage.setItem('logged', 'true');
-      sessionStorage.setItem('username', username);
-      welcomeMessage.textContent = `Olá, ${username}!`;
-      const loginMessage = document.getElementById("loginMessage");
-      loginMessage.style.visibility= 'visible';
-      const login = document.getElementById('loginButton');
-      login.style.visibility='hidden';
-      console.log('user ' + username + ' logged');
+      if(checkUsernameExists(username)){
+        sessionStorage.setItem('logged', 'true');
+        sessionStorage.setItem('username', username);
+        welcomeMessage.textContent = `Olá, ${username}!`;
+        const loginMessage = document.getElementById("loginMessage");
+        loginMessage.style.visibility= 'visible';
+        const login = document.getElementById('loginButton');
+        login.style.visibility='hidden';
+        console.log('user ' + username + ' logged');
+      }
+      else {
+        alert(username + " não encontrado. Por favor registe-se");
+      }
     }
 
 
@@ -83,6 +102,7 @@ w3.includeHTML(() =>  {
     const password = document.getElementById('new-password').value;
     const passwordConfirm = document.getElementById('new-passwordConfirm').value;
     const email = document.getElementById('new-email').value;
+    
 
     if(nome.trim() === ""){
       alert("O nome é um campo de preenchimento obrigatório");
@@ -90,7 +110,7 @@ w3.includeHTML(() =>  {
     else if(username.trim() === ""){
       alert("O username é de preenchimento obrigatório");
     }
-    else if(checkUsernameTaken(username)){
+    else if(checkUsernameExists(username)){
       alert("Esse username já existe");
     }
     else if(password.trim() === ""){
@@ -101,7 +121,7 @@ w3.includeHTML(() =>  {
     }
     else if(!comparePasswords(password, passwordConfirm)){
       console.log("deu falso");
-      alert("As passwords que escolheu não correspondem");
+      alert("As passwords não correspondem");
     }
     else if(email.trim() === ""){
       console.log("deu verdadeiro");
@@ -119,6 +139,7 @@ w3.includeHTML(() =>  {
         }
         users.push(username);
         localStorage.setItem('users', JSON.stringify(users));
+        login(username);
         form.reset();
         modalRegister.style.display = "none";
       }
@@ -134,16 +155,17 @@ w3.includeHTML(() =>  {
     }
   }
 
-  function checkUsernameTaken(username){
-    const users = localStorage.getItem('users');
-    user = JSON.stringify(users);
-    users.array.forEach(user => {
-      if(username.localeCompare(user)){
-        return true;
-      }
-    });
-    return false;
-  };
+  function checkUsernameExists(username){
+    console.log("a correr comparação de " + username);
+    let users = localStorage.getItem('users');
+    users = JSON.parse(users);
+    if(users.includes(username)){
+      return true;
+    }
+    else{
+      return false;
+    }
+  }
 
   // Função para alternar a exibição do aside
     function toggleAside() {
@@ -165,7 +187,25 @@ w3.includeHTML(() =>  {
     modalRegister.style.display = "none";
   }; 
 
+  
+
 });
+
+  //vê se pode ir para main
+  function goToMain(){
+    let logged = sessionStorage.getItem("logged");
+    // Verifica se 'logged' existe no localStorage
+    if (logged) {
+        logged = JSON.parse(logged);
+        console.log(logged);
+    }
+    if(logged){
+      window.location.href = "main.html";
+    }
+    else {
+      alert("Terá de fazer login para aceder à página de produtos");
+    }
+  }
 
   function showPassword() {
     console.log("start");
